[Platform] Changes to allow multiple Helm releases in one namespace
Adds a new universe configuration key HELM_NAMING_STYLE, the new
naming style is enabled for all the YugabyteDB installations which are
newer than 2.7.2.0-b113, as this version has Helm chart related
changes required for this naming to work.

Make changes to Create, Update, Edit flows, backup and healthchecks
scripts, metrics accordingly.

When we try to create multiple Helm releases in one namespace, we try
to apply the same pull secret in same namespace in parallel. There is
a race condition with `kubectl apply` on the client side (kubectl
side), it fails with an error saying the secret already exist. This is
fixed by using server side apply, the flag is experimental in 1.15 but
from Kubernetes 1.16 the feature is marked as beta and hence available
everywhere.

Signed-off-by: Bhavin Gandhi <bhavin7392@gmail.com>